[property-svc] test mongo adapter: property.Repository
